backend/news_service/requirements.txt
annotated-types
anyio
bcrypt
certifi
click
colorama
dnspython
email_validator
fastapi
fastapi-cli
h11
httpcore
httptools
httpx
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
passlib
pydantic
pydantic_core
Pygments
pymongo
python-dotenv
python-multipart
PyYAML
rich
rich-toolkit
shellingham
sniffio
starlette
typer
typing-inspection
typing_extensions
uvicorn
watchfiles
websockets
